requirements.txt
streamlit>=1.28.0
python-docx>=1.1.0
python-pptx>=0.6.21
openpyxl>=3.1.0
